requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.4.
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
alembic==1.11.1
    # via -r requirements.in
async-timeout==4.0.2
    # via aiohttp
asyncpg==0.27.0
    # via -r requirements.in
attrs==23.1.0
    # via aiohttp
charset-normalizer==3.1.0
    # via aiohttp
exceptiongroup==1.1.1
    # via pytest
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
greenlet==2.0.2
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via yarl
importlib-metadata==6.6.0
    # via alembic
importlib-resources==5.12.0
    # via alembic
iniconfig==2.0.0
    # via pytest
mako==1.2.4
    # via alembic
markupsafe==2.1.2
    # via mako
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
packaging==23.1
    # via pytest
pluggy==1.0.0
    # via pytest
psycopg2-binary==2.9.6
    # via -r requirements.in
pyjwt==2.6.0
    # via -r requirements.in
pytest==7.3.1
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
sqlalchemy[asyncio]==2.0.15
    # via
    #   -r requirements.in
    #   alembic
tomli==2.0.1
    # via pytest
typing-extensions==4.6.2
    # via
    #   alembic
    #   sqlalchemy
yarl==1.9.2
    # via aiohttp
zipp==3.15.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
